Issue #3374395: The file assets://css was not deleted because it does not exist.

See merge request project/drupal!4378

+++ core/lib/Drupal/Core/Asset/CssCollectionOptimizerLazy.php
@@ -143,7 +143,9 @@ class CssCollectionOptimizerLazy implements AssetCollectionGroupOptimizerInterfa
    */
   public function deleteAll() {
     $this->state->delete('drupal_css_cache_files');
-    $this->fileSystem->deleteRecursive('assets://css');
+    if (is_dir('assets://css')) {
+      $this->fileSystem->deleteRecursive('assets://css');
+    }
   }
 
   /**
